chore(thermal): better code organization, add some design note
