+++ antarest/study/business/areas/thermal_management.py
@@ -29,11 +29,11 @@ from antarest.study.business.model.thermal_model import (
     ThermalCluster,
     ThermalClusterCreation,
     ThermalClusterUpdate,
-    parse_thermal_cluster,
 )
 from antarest.study.business.study_interface import StudyInterface
 from antarest.study.model import STUDY_VERSION_8_7
 from antarest.study.storage.rawstudy.model.filesystem.config.identifier import transform_name_to_id
+from antarest.study.storage.rawstudy.model.filesystem.config.thermal import parse_thermal_cluster
 from antarest.study.storage.variantstudy.model.command.create_cluster import CreateCluster
 from antarest.study.storage.variantstudy.model.command.remove_cluster import RemoveCluster
 from antarest.study.storage.variantstudy.model.command.replace_matrix import ReplaceMatrix
